Make e2e tests run on Sauce Labs.
