+++ package.json
@@ -55,6 +55,7 @@
     "retry": "0.10.1",
     "roboter": "1.0.4",
     "runfork": "0.3.1",
+    "sauce-connect-launcher": "1.2.4",
     "selenium-standalone": "6.15.0",
     "selenium-webdriver": "4.0.0-alpha.1",
     "shelljs": "0.7.8",
